separate picmi default and yee solver into two files
