+++ lib/python/picongpu/pypicongpu/field_solver/__init__.py
@@ -1,0 +1,2 @@
+from .DefaultSolver import Solver as Solver
+from .Yee import YeeSolver as YeeSolver
